licences move resource
