+++ internal/meraki/organizations/licences/move/resource.go
@@ -1,4 +1,4 @@
-package organizations
+package move
 
 import (
 	"context"
@@ -16,20 +16,20 @@ import (
 )
 
 var (
-	_ resource.Resource                = &LicenseResource{} // Terraform resource interface
-	_ resource.ResourceWithConfigure   = &LicenseResource{} // Interface for resources with configuration methods
-	_ resource.ResourceWithImportState = &LicenseResource{} // Interface for resources with import state functionality
+	_ resource.Resource                = &Resource{} // Terraform resource interface
+	_ resource.ResourceWithConfigure   = &Resource{} // Interface for resources with configuration methods
+	_ resource.ResourceWithImportState = &Resource{} // Interface for resources with import state functionality
 )
 
-func NewOrganizationsLicenseResource() resource.Resource {
-	return &LicenseResource{}
+func NewResource() resource.Resource {
+	return &Resource{}
 }
 
-type LicenseResource struct {
+type Resource struct {
 	client *openApiClient.APIClient // APIClient instance for making API requests
 }
 
-type LicenseResourceModel struct {
+type resourceModel struct {
 	Id                 jsontypes.String   `tfsdk:"id" json:"-"`
 	OrganizationId     jsontypes.String   `tfsdk:"organization_id"`
 	DestOrganizationId jsontypes.String   `tfsdk:"dest_organization_id"`
@@ -38,7 +38,7 @@ type LicenseResourceModel struct {
 
 // Metadata provides a way to define information about the resource.
 // This method is called by the framework to retrieve metadata about the resource.
-func (r *LicenseResource) Metadata(ctx context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse) {
+func (r *Resource) Metadata(ctx context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse) {
 
 	// The TypeName attribute is important as it provides the user-friendly name for the resource/data source.
 	// This is the name users will use to reference the resource/data source and it's also used in the acceptance tests.
@@ -47,7 +47,7 @@ func (r *LicenseResource) Metadata(ctx context.Context, req resource.MetadataReq
 
 // Schema provides a way to define the structure of the resource data.
 // It is called by the framework to get the schema of the resource.
-func (r *LicenseResource) Schema(ctx context.Context, req resource.SchemaRequest, resp *resource.SchemaResponse) {
+func (r *Resource) Schema(ctx context.Context, req resource.SchemaRequest, resp *resource.SchemaResponse) {
 
 	// The Schema object defines the structure of the resource.
 	resp.Schema = schema.Schema{
@@ -85,7 +85,7 @@ func (r *LicenseResource) Schema(ctx context.Context, req resource.SchemaRequest
 
 // Configure is a method of the Resource interface that Terraform calls to provide the configured provider instance to the resource.
 // It passes the ResourceData that's been stored by the provider's ConfigureFunc.
-func (r *LicenseResource) Configure(ctx context.Context, req resource.ConfigureRequest, resp *resource.ConfigureResponse) {
+func (r *Resource) Configure(ctx context.Context, req resource.ConfigureRequest, resp *resource.ConfigureResponse) {
 
 	// The provider must be properly configured before it can be used.
 	if req.ProviderData == nil {
@@ -113,9 +113,9 @@ func (r *LicenseResource) Configure(ctx context.Context, req resource.ConfigureR
 // Create method is responsible for creating a new resource.
 // It takes a CreateRequest containing the planned state of the new resource and returns a CreateResponse
 // with the final state of the new resource or an error.
-func (r *LicenseResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
+func (r *Resource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
 
-	var data *LicenseResourceModel
+	var data *resourceModel
 
 	// Unmarshal the plan data into the internal data model struct.
 	resp.Diagnostics.Append(req.Plan.Get(ctx, &data)...)
@@ -181,8 +181,8 @@ func (r *LicenseResource) Create(ctx context.Context, req resource.CreateRequest
 
 // Read method is responsible for reading an existing resource's state.
 // It takes a ReadRequest and returns a ReadResponse with the current state of the resource or an error.
-func (r *LicenseResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
-	var data *LicenseResourceModel
+func (r *Resource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
+	var data *resourceModel
 
 	// Read Terraform configuration data into the model
 	resp.Diagnostics.Append(req.State.Get(ctx, &data)...)
@@ -201,9 +201,9 @@ func (r *LicenseResource) Read(ctx context.Context, req resource.ReadRequest, re
 
 // Update function is responsible for updating the state of an existing resource.
 // It uses an UpdateRequest and responds with an UpdateResponse which contains the updated state of the resource or an error.
-func (r *LicenseResource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
+func (r *Resource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
 
-	var data *LicenseResourceModel
+	var data *resourceModel
 
 	resp.Diagnostics.Append(req.Plan.Get(ctx, &data)...)
 
@@ -221,9 +221,9 @@ func (r *LicenseResource) Update(ctx context.Context, req resource.UpdateRequest
 
 // Delete function is responsible for deleting a resource.
 // It uses a DeleteRequest and responds with a DeleteResponse which contains the updated state of the resource or an error.
-func (r *LicenseResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
+func (r *Resource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
 
-	var data *LicenseResourceModel
+	var data *resourceModel
 
 	resp.Diagnostics.Append(req.State.Get(ctx, &data)...)
 
@@ -241,7 +241,7 @@ func (r *LicenseResource) Delete(ctx context.Context, req resource.DeleteRequest
 // ImportState function is used to import an existing resource into Terraform.
 // The function expects an ImportStateRequest and responds with an ImportStateResponse which contains
 // the new state of the resource or an error.
-func (r *LicenseResource) ImportState(ctx context.Context, req resource.ImportStateRequest, resp *resource.ImportStateResponse) {
+func (r *Resource) ImportState(ctx context.Context, req resource.ImportStateRequest, resp *resource.ImportStateResponse) {
 
 	// Pass through the ID directly from the ImportStateRequest to the ImportStateResponse
 	resource.ImportStatePassthroughID(ctx, path.Root("id"), req, resp)
